move standard processing functions to isce_util.py
Continuing clean up; move some standard processing functions to isce_util.py
