+++ mintpy/utils/isce_utils.py
@@ -570,7 +570,7 @@ def read_tops_baseline(baseline_file):
 def read_stripmap_baseline(baseline_file):
     """Read baseline file generated by ISCE/stripmapStack processor.
 
-    Example: 
+    Example:
     baselines/20200111_20200125.txt
         PERP_BASELINE_BOTTOM 173.97914535263297
         PERP_BASELINE_TOP 174.05612879066618
@@ -917,3 +917,110 @@ def get_sensing_datetime_list(proj_dir, date_list=None):
 
     return sensingMid, sensingStart, sensingStop
 
+############################## Standard Processing ###########################################
+
+def gaussian_kernel(Sx, Sy, sig_x, sig_y):
+    if np.mod(Sx,2) == 0:
+        Sx = Sx + 1
+
+    if np.mod(Sy,2) ==0:
+            Sy = Sy + 1
+
+    x,y = np.meshgrid(np.arange(Sx),np.arange(Sy))
+    x = x + 1
+    y = y + 1
+    x0 = (Sx+1)/2
+    y0 = (Sy+1)/2
+    fx = ((x-x0)**2.)/(2.*sig_x**2.)
+    fy = ((y-y0)**2.)/(2.*sig_y**2.)
+    k = np.exp(-1.0*(fx+fy))
+    a = 1./np.sum(k)
+    k = a*k
+    return k
+
+def convolve(data, kernel):
+    import cv2
+    R = cv2.filter2D(data.real,-1,kernel)
+    Im = cv2.filter2D(data.imag,-1,kernel)
+
+    return R + 1J*Im
+
+def write_xml(filename,width,length,bands,dataType,scheme):
+    import isce
+    import isceobj
+    img=isceobj.createImage()
+    img.setFilename(filename)
+    img.setWidth(width)
+    img.setLength(length)
+    img.setAccessMode('Read')
+    img.bands=bands
+    img.dataType=dataType
+    img.scheme = scheme
+    img.renderHdr()
+    img.renderVRT()
+    return
+
+def estCoherence(outfile, corfile):
+    import isce
+    import isceobj
+    from mroipac.icu.Icu import Icu
+
+    #Create phase sigma correlation file here
+    filtImage = isceobj.createIntImage()
+    filtImage.load( outfile + '.xml')
+    filtImage.setAccessMode('read')
+    filtImage.createImage()
+
+    phsigImage = isceobj.createImage()
+    phsigImage.dataType='FLOAT'
+    phsigImage.bands = 1
+    phsigImage.setWidth(filtImage.getWidth())
+    phsigImage.setFilename(corfile)
+    phsigImage.setAccessMode('write')
+    phsigImage.createImage()
+
+
+    icuObj = Icu(name='sentinel_filter_icu')
+    icuObj.configure()
+    icuObj.unwrappingFlag = False
+    icuObj.useAmplitudeFlag = False
+    #icuObj.correlationType = 'NOSLOPE'
+
+    icuObj.icu(intImage = filtImage,  phsigImage=phsigImage)
+    phsigImage.renderHdr()
+
+    filtImage.finalizeImage()
+    phsigImage.finalizeImage()
+
+    return
+
+def unwrap_snaphu(intfile,corfile,unwfile,length, width):
+    import isce
+    from contrib.Snaphu.Snaphu import Snaphu
+
+    altitude = 800000.0
+    earthRadius = 6371000.0
+    wavelength = 0.056
+
+    snp = Snaphu()
+    snp.setInitOnly(False)
+    snp.setInput(intfile)
+    snp.setOutput(unwfile)
+    snp.setWidth(width)
+    snp.setCostMode('DEFO')
+    snp.setEarthRadius(earthRadius)
+    snp.setWavelength(wavelength)
+    snp.setAltitude(altitude)
+    snp.setCorrfile(corfile)
+    snp.setInitMethod('MST')
+   # snp.setCorrLooks(corrLooks)
+    snp.setMaxComponents(100)
+    snp.setDefoMaxCycles(2.0)
+    snp.setRangeLooks(80)
+    snp.setAzimuthLooks(20)
+    snp.setCorFileFormat('FLOAT_DATA')
+    snp.prepare()
+    snp.unwrap()
+
+    return
+    write_xml(unwfile, width, length, 2 , "FLOAT",'BIL')
